requirements.txt
bme680
influxdb-client
python-dotenv